test: cover decoupled producer/consumer streaming
